# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "User Login",
                "description": "Login User Here",
                "operationId": "authLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "email",
                                        "type": "string",
                                        "example": "test@domain.com"
                                    },
                                    "password": {
                                        "description": "password",
                                        "type": "string",
                                        "example": "12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/agent/notification": {
            "post": {
                "tags": [
                    "Agent"
                ],
                "summary": "Store new notification",
                "description": "Returns notification data",
                "operationId": "storeNotification",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "clientId": {
                                        "description": "clientId",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "channel": {
                                        "description": "channel",
                                        "type": "string",
                                        "example": "sms"
                                    },
                                    "content": {
                                        "description": "content",
                                        "type": "string",
                                        "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit.."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Notifications"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/agent/client/{id}": {
            "get": {
                "tags": [
                    "Agent"
                ],
                "summary": "Get client information by id",
                "description": "Returns client data",
                "operationId": "getClientById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "client id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Clients"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/agent/allClients": {
            "get": {
                "tags": [
                    "Agent"
                ],
                "summary": "Get all clients",
                "description": "Get paginated list of clients",
                "operationId": "getClientsList_",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Clients"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/agent/notification/{id}": {
            "get": {
                "tags": [
                    "Agent"
                ],
                "summary": "Get notification information",
                "description": "Return notification by id",
                "operationId": "getNotificationById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Notification id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Notifications"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/agent/filterNotification": {
            "get": {
                "tags": [
                    "Agent"
                ],
                "summary": "Get paginated list of notifications with filter",
                "description": "get paginated list of notifications, with possibility to filter notifications by client",
                "operationId": "getPaginatedNotification_",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "clientId",
                        "in": "query",
                        "description": "Filter client Id",
                        "schema": {
                            "type": "object"
                        },
                        "example": {
                            "clientId": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Notifications"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Clients": {
                "title": "Clients",
                "description": "Clients model",
                "properties": {
                    "firstName": {
                        "title": "firstname",
                        "description": "Customer name",
                        "type": "string",
                        "format": "string"
                    },
                    "lastName": {
                        "title": "lastName",
                        "description": "Customer lastName",
                        "type": "string",
                        "format": "string"
                    },
                    "email": {
                        "title": "email",
                        "description": "Customer email",
                        "type": "string",
                        "format": "string"
                    },
                    "phoneNumber": {
                        "title": "phonenumbar",
                        "description": "Customer phonenumbar",
                        "type": "string",
                        "format": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Clients"
                }
            },
            "Notifications": {
                "title": "Notifications",
                "description": "Notifications model",
                "properties": {
                    "clientId": {
                        "title": "clientId",
                        "description": "Customer id",
                        "type": "string",
                        "format": "integer"
                    },
                    "channel": {
                        "title": "channel",
                        "description": "message channel",
                        "type": "string",
                        "format": "string"
                    },
                    "content": {
                        "title": "content",
                        "description": "Message text",
                        "type": "string",
                        "format": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Notifications"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}